Update paths. Move all code into src/ folder.
Contents of src folder contains:

* client/ - Used for publicly built client code, mainly html, css and js.
* lib/ - Shared code between client and server.
* server/ - Node.js server and libraries.
