+++ src/client/img/crosshair.png
