+++ src/client/img/player.png
